refactor(worker): inline temporal module to root level
Move all files from the temporal/ subdirectory to the crate root:
- temporal/types.rs → temporal_types.rs
- temporal/activities.rs → temporal_activities.rs
- temporal/client.rs → temporal_client.rs
- temporal/runner.rs → temporal_runner.rs
- temporal/workflows/agent_run.rs → session_workflow.rs

This flattens the worker crate structure by removing one level of
nesting. All imports updated throughout the crate.
