+++ crates/everruns-worker/src/temporal_client.rs
@@ -21,8 +21,7 @@ use temporal_sdk_core::{
 };
 
 use crate::runner::RunnerConfig;
-
-use super::types::{workflow_names, SessionWorkflowInput};
+use crate::temporal_types::{workflow_names, SessionWorkflowInput};
 
 /// Client for interacting with Temporal server
 /// Used by the API to start workflows
