+++ crates/everruns-worker/src/temporal_runner.rs
@@ -15,10 +15,9 @@ use tracing::{error, info};
 use uuid::Uuid;
 
 use crate::runner::{AgentRunner, RunnerConfig};
+use crate::temporal_client::TemporalClient;
+use crate::temporal_types::SessionWorkflowInput;
 use crate::worker::TemporalWorker;
-
-use super::client::TemporalClient;
-use super::types::SessionWorkflowInput;
 
 /// Temporal-based session runner using true Temporal workflows
 ///
