+++ crates/everruns-worker/src/worker.rs
@@ -30,13 +30,14 @@ use tokio::task::JoinHandle;
 use tracing::{debug, error, info, warn};
 
 use crate::runner::RunnerConfig;
-use crate::temporal::activities::{
+use crate::temporal_activities::{
     call_llm_activity, execute_tools_activity, load_agent_activity, load_messages_activity,
     persist_event_activity, save_message_activity, update_status_activity, ActivityContext,
 };
-use crate::temporal::client::TemporalWorkerCore;
-use crate::temporal::types::*;
-use crate::temporal::workflows::{Workflow, WorkflowAction, WorkflowRegistry};
+use crate::temporal_client::TemporalWorkerCore;
+use crate::temporal_types::*;
+use crate::workflow_registry::WorkflowRegistry;
+use crate::workflow_traits::Workflow;
 
 /// Temporal worker that processes workflow and activity tasks
 pub struct TemporalWorker {
